Restore lazy evaluation of fallible CASE  (#15390)
* Restore lazy evaluation of fallible CASE

Commit 0f4b8b136ceb9132fd6b6595bd6a09a09707f5d9 introduced a new
optimized evaluation mode for CASE expression with exactly one WHEN-THEN
clause and ELSE clause being present. Apparently, the new mode did not
take into account expensive or fallible expressions, as it
unconditionally evaluated both branches. This is definitely incorrect
for fallible expressions. For these, fallback to the original execution
mode.

* Workaround for running sqllogictests in RustRover